Fixed appId to be int64 instead of int

+++ types/app_env.go
@@ -2,8 +2,8 @@ package types
 
 type AppEnv struct {
 	IdModel
-	AppId   int    `json:"appId"`
-	EnvId   int    `json:"envId"`
+	AppId   int64  `json:"appId"`
+	EnvId   int64  `json:"envId"`
 	Version string `json:"version"`
 
 	App *Application `json:"app,omitempty"`
